+++ lib/plausible_api/stats/timeseries.rb
@@ -2,19 +2,15 @@
 
 module PlausibleApi
   module Stats
-    class Timeseries
+    class Timeseries < Base
       def initialize(options = {})
         @period   = options[:period] || '30d'
         @filters  = options[:filters]
         @interval = options[:interval]
       end
 
-      def request_url
-        url = "/api/v1/stats/timeseries?site_id=$SITE_ID"
-        url += "&period=#{@period}" if @period
-        url += "&filters=#{CGI.escape(@filters)}" if @filters
-        url += "&interval=#{@interval}" if @interval
-        url
+      def request_url_base
+        "/api/v1/stats/timeseries?site_id=$SITE_ID"
       end
 
       def parse_response(body)
