+++ lib/plausible_api/version.rb
@@ -1,3 +1,3 @@
 module PlausibleApi
-  VERSION = "0.1.4"
+  VERSION = "0.1.5"
 end
